SI-7020 Deterministic warnings for pattern matcher, take 2
The previous swing at determinism, ebb01e05cbe4, made decent
contact but apparently didn't hit it out of the park. The test
wavered every hundred or so runs, as witnessed occasionally in
nightly builds or pull request validation.

I setup a test to run neg/7020.scala a few hundred times, and
could trigger the failure reliably.

I then swept through the pattern matcher in search of HashMap and
HashSet creation, and changed them all to the Linked variety.
The results of that are published in retronym#ticket/7020-3 [1].

This commit represents the careful whittling down of that patch
to the minimal change required to exhibit determinism.

[1] https://github.com/retronym/scala/compare/ticket/7020-3