M6 modules, partest depends on scala as provided.
The new partest now marks its scala dependencies as provided.
This is much more robust than before, and it breaks ugly cycles.

Updated the build for partest-extras to provide the required
scala-compiler dependency.

Also, upgrade pax-url-aether to 1.6.0.

Since our jenkins uses mirrors with passwords,
we needed a fix for https://ops4j1.jira.com/browse/PAXURL-217
in order to run osgi.test on jenkins, now that we use maven more.
We didn't hit this bug before because we were using a standard
location for the maven local repository, but that causes problems
with concurrent jenkins jobs accessing it.

So, I added the following to `~/.m2/settings`,
and then ran into the issue described above.

Also removed some non-essential uses of the extra-repo (slows stuff down),
and added a comment of how to make the osgi tests look in a certain
local m2 repo (for /some/ operations...)

```
<!--
  Maven repos are not safe for concurrent access, thus give each job its own.
  The WORKSPACE env var is set by jenkins to the path of the current job's workspace.
  This settings file will cause ant/maven to break if WORKSPACE is not set (to a writeable directory).
  I don't know how to encode a fall back, but I assume it's not needed as this is ~jenkins/.m2/settings.xml
-->
<localRepository>${WORKSPACE}/m2repo</localRepository>
```