Fixes broken 'Symbol-handling code in CleanUp
Looks like the transformation did never happen because the pattern
failed to match. Why did the pattern fail to match? Because the
Symbol.apply we see in the tree claims to be a method while
Symbol_apply defined in Definitions wants to be a value.

Why this happens remains a mystery to me.